+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [6.5.1](https://github.com/ant-design/ant-design-cli/compare/v6.5.0...v6.5.1) (2026-07-13)
+
+- Update antd metadata ([v6@6.5.1](https://github.com/ant-design/ant-design-cli/compare/v6.5.0...v6.5.1#diff-ebaa5874f72b5c0a62edf9d98d6ae55fffc16dc881ade7a697e589c8614c7436))
+
+
 ## Unreleased
 
 ### Bug Fixes